+++ packages/screen_retriever/test/screen_retriever_test.dart
@@ -16,6 +16,10 @@ class MockScreenRetrieverPlatform
     with MockPlatformInterfaceMixin
     implements ScreenRetrieverPlatform {
   @override
+  Stream<Map<Object?, Object?>> get onScreenEventReceiver =>
+      throw UnimplementedError();
+
+  @override
   Future<Offset> getCursorScreenPoint() {
     return Future(() => const Offset(10.0, 10.0));
   }