+++ packages/screen_retriever_platform_interface/lib/src/screen_listener.dart
@@ -1,5 +1,5 @@
 /// Interface for listening to screen events.
-abstract class ScreenListener {
+abstract mixin class ScreenListener {
   /// Called when a screen event occurs.
   void onScreenEvent(String eventName) {}
 }